fix: profile social components now get user as prop
