+++ app/screens/profile/Social.tsx
@@ -1,49 +1,47 @@
 import React from "react";
-import { TouchableOpacity, View } from "react-native";
+import { View, StyleSheet } from "react-native";
 import { shadow } from "../../themes/theme";
-import { Feather } from "@expo/vector-icons";
-import { loadInBrowser } from "../../utils";
 import { useConfig } from "../../hooks";
+import { user } from "../../hooks/types";
+import SocialButton from "./SocialButton";
 
-interface ISocialButton {
-  uri: string;
-  icon: string;
+interface ISocial {
+  user: user;
 }
 
-export function SocialButton({ icon, uri }: ISocialButton): JSX.Element {
-  const { config } = useConfig();
-  const { theme } = config;
-  return (
-    <TouchableOpacity
-      onPress={() => loadInBrowser(uri)}
-      style={{ flex: 1, justifyContent: "center", alignItems: "center" }}
-    >
-      <Feather
-        name={icon}
-        size={20}
-        color={theme.colors.primary}
-        style={{ marginRight: 10 }}
-      />
-    </TouchableOpacity>
-  );
-}
-
-export function Social(): JSX.Element {
+export function Social({ user }: ISocial): JSX.Element {
   const { config } = useConfig();
   const { theme } = config;
   return (
     <View
       style={[
         {
-          height: 50,
           backgroundColor: theme.colors.foreground,
-          borderRadius: 10,
-          flexDirection: "row",
-          alignItems: "center",
-          paddingHorizontal: 10,
         },
         shadow,
+        styles.container,
       ]}
-    ></View>
+    >
+      {user.twitch ? (
+        <SocialButton icon="twitch" uri={user.twitch.uri} />
+      ) : null}
+      {user.youtube ? (
+        <SocialButton icon="youtube" uri={user.youtube.uri} />
+      ) : null}
+      {user.twitter ? (
+        <SocialButton icon="twitter" uri={user.twitter.uri} />
+      ) : null}
+      {user.hitbox ? <SocialButton icon="globe" uri={user.hitbox.uri} /> : null}
+    </View>
   );
 }
+
+const styles = StyleSheet.create({
+  container: {
+    height: 50,
+    borderRadius: 10,
+    flexDirection: "row",
+    alignItems: "center",
+    paddingHorizontal: 10,
+  },
+});
